split detector connections and segment/session generation
